third_party/WebKit/LayoutTests/http/tests/htmlimports/reject-content-disposition.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link id="target" rel="import" href="resources/content-disposition.cgi">
</head>
<body>
<script>
test(function() {
  var target = document.getElementById("target");
  assert_true(target.import === null);
},
"Check if Content-Disposition prevents imports from loading",
{
  "help": ["http://w3c.github.io/webcomponents/spec/imports/#fetching-import"],
});
</script>
</body>
</html>
